requirements.txt
install==1.3.5
numpy==1.26.4
pandas==2.2.0
pillow==10.2.0
pyarrow==15.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tzdata==2023.4
